Implement the game ending logic
An RPS game will now end when either the player or cpu reaches 5 wins.
Every time the user clicks any of the buttons, the game function checks
if either the player or the cpu currently has 5 wins and if so it sets
a flag to make the next button click function as a game reset.